Creación del ChatBot Component
